requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
email-validator==2.2.0
python-multipart==0.0.20
python-dotenv==1.1.1
